Update worker organization
